Module sorting_visualization::app

The App struct and some constants.

Structs

App
This struct contains all rendering, updating and input handling logic.

Constants

ACCESSED_VALUE_COLOR
Color of the values that have been recently accessed.

ACCESSED_VALUE_TIMEOUT
Time in seconds after which array accesses are removed.

BACKGROUND_COLOR
Color that is used to clear the window before drawing.

SPEED_FACTOR
Factor for increasing / decreasing speed of sorting array

STATUS_TEXT_FONT_SIZE
Font size of the status text in pixels.

STATUS_TEXT_MARGIN
Margins between the status text and window borders.

VALUE_COLOR
Color of rectangles that represent the array values.