Struct sorting_visualization::app::App

pub struct App {
    state: SharedState,
    algorithm_thread: JoinHandle<()>,
}

This struct contains all rendering, updating and input handling logic.

Fields
state: SharedStatealgorithm_thread: JoinHandle<()>
Implementations

impl App

pub fn init(
algorithm: Box<dyn Algorithm + Send>,
array: Vec<u32>,
speed: f64
) -> Self

Creates a new app (with a state constructed from the given array) and starts an algorithm thread. This function is called init instead of new because it has side effects.

pub fn render(
&mut self,
args: RenderArgs,
gl: &mut GlGraphics,
glyphs: &mut GlyphCache<'_>
)

Draws the current state.

pub fn update(&mut self, args: UpdateArgs)

Advances the state by a given amount of time.

pub fn button(&mut self, args: ButtonArgs)

Handles user input and updates the state.

Controls

Key	Action

Space	pause/resume

↑	2x faster

↓	2x slower