Module sorting_visualization::cli

Command-line interface and command-line argument parsing. Uses clap under the hood.

Structs

Options
Contains all options that can be provided by a user using the CLI.

Enums

Order
Order of elements in the array.

Constants

ALGORITHM_ARG
Internal name of the algorithm argument which is used to get its value.

LENGTH_OPT
Internal name of the --length/-l option which is used to get its value.

ORDER_OPT
Internal name of the --order/-o option which is used to get its value.

Functions

parse_options
Parses command-line arguments into Options. If the help/version is printed or an error occurs, it will be displayed to the user and the process will exit.