Module sorting_visualization::state

Structs and type definitions that describe the app state.

Structs

SharedState
A wrapper around State that can be safely shared between threads.

State
Contains the state of the whole app.

Constants

NO_ARRAY_ACCESS
A constant that means "there's no array access here". See documentation of State.array_accesses to understand why this constant is a negative number.