Trait sorting_visualization::algorithms::Algorithm

pub trait Algorithm {
    fn sort(&self, array: Array);

    fn name(&self) -> String;
}

The general trait for all sorting algorithms.

Required methods

fn sort(&self, array: Array)

Sorts a given array. This method is called in a so called "algorithm thread".

fn name(&self) -> String

Returns the name of the algorithm that will be displayed to the user. Returned value is an owned String so it can be generated at runtime.

Loading content...